app/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
Werkzeug==2.2.3
